Add watermask support and rw light

+++ forge/tests/test_terrain.py
@@ -3,15 +3,20 @@
 import unittest
 import os
 from forge.terrain import TerrainTile
+from forge.lib.global_geodetic import GlobalGeodetic
 
 
 class TestTerrainTile(unittest.TestCase):
 
     def setUp(self):
-        self.tmpfile = 'temp.terrain'
+        self.tmpfile = '.tmp/temp.terrain'
+        self.tmpfile2 = '.tmp/temp2.terrain'
 
     def tearDown(self):
-        os.remove(self.tmpfile)
+        if os.path.exists(self.tmpfile):
+            os.remove(self.tmpfile)
+        if os.path.exists(self.tmpfile2):
+            os.remove(self.tmpfile2)
 
     def testReaderWriter(self):
         '''
@@ -76,3 +81,44 @@ class TestTerrainTile(unittest.TestCase):
         self.assertEqual(len(ter.northI), len(ter2.northI))
         for i, v in enumerate(ter.northI):
             self.assertEqual(v, ter2.northI[i], i)
+
+    def testExtensionsReader(self):
+        z = 10
+        y = 1563
+        x = 590
+        geodetic = GlobalGeodetic(True)
+
+        ter = TerrainTile()
+        [minx, miny, maxx, maxy] = geodetic.TileBounds(x, y, z)
+        ter.fromFile('forge/data/quantized-mesh/%s_%s_%s.terrain' % (z, y, x),
+            minx, miny, maxx, maxy, hasLighting=True, hasWatermask=True)
+
+        # check indices
+        self.failUnless(len(ter.indices) > 0)
+
+        # check edges
+        self.failUnless(len(ter.westI) > 0)
+        self.failUnless(len(ter.southI) > 0)
+        self.failUnless(len(ter.eastI) > 0)
+        self.failUnless(len(ter.northI) > 0)
+
+        # check extensions
+        self.assertEqual(len(ter.watermask), 1)
+        self.assertEqual(len(ter.watermask[0]), 1)
+        # Water only -> 255
+        self.assertEqual(ter.watermask[0][0], 255)
+        ter.toFile(self.tmpfile2)
+
+        ter2 = TerrainTile()
+        ter2.fromFile(self.tmpfile2,
+            minx, miny, maxx, maxy, hasLighting=True, hasWatermask=True)
+
+        self.assertEqual(len(ter.watermask), len(ter2.watermask))
+        self.assertEqual(len(ter.watermask[0]), len(ter2.watermask[0]))
+
+        sign = lambda a: 1 if a > 0 else -1 if a < 0 else 0
+        for i in range(0, len(ter.vLight)):
+            for j in range(0, 3):
+                # We cannot have an exact equality with successive oct encoding and decoding
+                # Thus we only check the sign
+                self.assertEqual(sign(ter.vLight[i][j]), sign(ter2.vLight[i][j]))
